requirements.txt
Django~=5.0.2
djangorestframework==3.15.2
requests==2.31.0
# psycopg2==2.9.9
python-dotenv==1.0.1
djangorestframework-simplejwt==4.7.2
